+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add labels to `database show` output.
+- In `router show` command, list static routes in the human readable output and add static route type field to all outputs.
 
 ## [3.10.0] - 2024-07-17